backend/requirements.txt
# Backend Dependencies - Django REST Framework
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
pandas==2.2.0
openpyxl==3.1.2
numpy==1.26.3
gunicorn==21.2.0
whitenoise==6.6.0
drf-yasg==1.21.7
pillow==10.2.0
celery==5.3.6
redis==5.0.1

